data/requirements.txt
fastapi
uvicorn
requests
pymysql
python-dotenv
confluent-kafka
pyspark==3.2.1
boto3
Pillow
hdfs